pod patches for English errors
Here are the patches for the pod files in Perl5.003, subversion 11.
The following corrections were made:

Spelling, grammar, jargon were fixed.
"It's" is the contraction of "it" and "is".
"Its" is the possessive form of "it".

only/just/simply/merely

Here's the example sentence that illustrates that "only" can be used in
almost any position in a sentence, and that each new sentence means
something different.  As a writer, your goal should be to communicate your
ideas unambiguously.

She kissed him on the lips.

Sentence				Meaning of "only"
a. Only she kissed him on the lips.	She and no one else
b. She only kissed him on the lips.	She didn't do anything else with his
					lips.
c. She kissed only him on the lips.	She didn't kiss anyone else on the
					lips.
d. She kissed him only on the lips.	not below or next to the lips
e. She kissed him on only the lips.	She didn't kiss him anywhere else.
f. She kissed him on the lips only.	ambiguous--the meaning could be any
					of a-e, depending on vocal stress
since/because
lend/loan
which/that

i.e. & ie => i.e.,
e.g. & eg => e.g.,

in order to => to
split infinitives
outer most => outermost
doubled words (to, the, be)
to able=> to be able

I hope this helps!

p5p-msgid: <24616.850167191@castor>